refactor verifier
