+++ verifier/client.go
